+++ examples/run/run.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 import sys
-sys.path.append("/home/user/path/to/adsorpyon/folder")
+sys.path.append("/home/user/path/to/src/folder")
 import main
 
 main.main()
